requirement.txt
Flask
Flask-JWT-Extended
psycopg2-binary 
bcrypt 
Flask-SQLAlchemy
